move net-status function to parent class
Add checks so ForceStopped controllers can’t be woken up by a net-event.
resolves nachocove/qa#884
